improve : improve performance

+++ config/src/lib.rs
@@ -37,9 +37,9 @@ pub struct StorageConfig {
     pub base_file_size: u64,
     pub compact_trigger: u32,
     pub max_compact_size: u64,
-    pub dio_max_resident: u64,
-    pub dio_max_non_resident: u64,
-    pub dio_page_len_scale: u64,
+    pub dio_max_resident: usize,
+    pub dio_max_non_resident: usize,
+    pub dio_page_len_scale: usize,
     pub strict_write: bool,
 }
 
@@ -64,13 +64,13 @@ impl StorageConfig {
             self.max_compact_size = size.parse::<u64>().unwrap();
         }
         if let Ok(size) = std::env::var("CNOSDB_STORAGE_DIO_MAX_RESIDENT") {
-            self.dio_max_resident = size.parse::<u64>().unwrap();
+            self.dio_max_resident = size.parse::<usize>().unwrap();
         }
         if let Ok(size) = std::env::var("CNOSDB_STORAGE_DIO_MAX_NON_RESIDENT") {
-            self.dio_max_non_resident = size.parse::<u64>().unwrap();
+            self.dio_max_non_resident = size.parse::<usize>().unwrap();
         }
         if let Ok(size) = std::env::var("CNOSDB_STORAGE_DIO_PAGE_LEN_SCALE") {
-            self.dio_page_len_scale = size.parse::<u64>().unwrap();
+            self.dio_page_len_scale = size.parse::<usize>().unwrap();
         }
         if let Ok(size) = std::env::var("CNOSDB_STORAGE_STRICT_WRITE") {
             self.strict_write = size.parse::<bool>().unwrap();
